Change Address model to be an interface. Will allow consuming projects to override this funcitonality more easily.
