+++ openid-connect-common/src/main/java/org/mitre/openid/connect/model/Address.java
@@ -18,219 +18,66 @@ package org.mitre.openid.connect.model;
 
 import java.io.Serializable;
 
-import javax.persistence.Basic;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="address")
-public class Address implements Serializable {
-
-	private static final long serialVersionUID = -1304880008685206811L;
-
-	private Long id;
-	private String formatted;
-	private String streetAddress;
-	private String locality;
-	private String region;
-	private String postalCode;
-	private String country;
+public interface Address extends Serializable {
 
 	/**
-	 * Empty constructor
+	 * @return the formatted address
 	 */
-	public Address() {
-
-	}
-
-	/**
-	 * @return the formatted address string
-	 */
-	@Basic
-	@Column(name = "formatted")
-	public String getFormatted() {
-		return formatted;
-	}
+	public String getFormatted();
+	
 	/**
 	 * @param formatted the formatted address to set
 	 */
-	public void setFormatted(String formatted) {
-		this.formatted = formatted;
-	}
+	public void setFormatted(String formatted);
+	
 	/**
 	 * @return the streetAddress
 	 */
-	@Basic
-	@Column(name="street_address")
-	public String getStreetAddress() {
-		return streetAddress;
-	}
+	public String getStreetAddress();
+	
 	/**
 	 * @param streetAddress the streetAddress to set
 	 */
-	public void setStreetAddress(String streetAddress) {
-		this.streetAddress = streetAddress;
-	}
+	public void setStreetAddress(String streetAddress);
+	
 	/**
 	 * @return the locality
 	 */
-	@Basic
-	@Column(name = "locality")
-	public String getLocality() {
-		return locality;
-	}
+	public String getLocality();
+	
 	/**
 	 * @param locality the locality to set
 	 */
-	public void setLocality(String locality) {
-		this.locality = locality;
-	}
+	public void setLocality(String locality);
+	
 	/**
 	 * @return the region
 	 */
-	@Basic
-	@Column(name = "region")
-	public String getRegion() {
-		return region;
-	}
+	public String getRegion();
+	
 	/**
 	 * @param region the region to set
 	 */
-	public void setRegion(String region) {
-		this.region = region;
-	}
+	public void setRegion(String region);
+	
 	/**
 	 * @return the postalCode
 	 */
-	@Basic
-	@Column(name="postal_code")
-	public String getPostalCode() {
-		return postalCode;
-	}
+	public String getPostalCode();
+	
 	/**
 	 * @param postalCode the postalCode to set
 	 */
-	public void setPostalCode(String postalCode) {
-		this.postalCode = postalCode;
-	}
+	public void setPostalCode(String postalCode);
+	
 	/**
 	 * @return the country
 	 */
-	@Basic
-	@Column(name = "country")
-	public String getCountry() {
-		return country;
-	}
+	public String getCountry();
+	
 	/**
 	 * @param country the country to set
 	 */
-	public void setCountry(String country) {
-		this.country = country;
-	}
-
-	/**
-	 * @return the id
-	 */
-	@Id
-	@GeneratedValue(strategy=GenerationType.IDENTITY)
-	@Column(name = "id")
-	public Long getId() {
-		return id;
-	}
-
-	/**
-	 * @param id the id to set
-	 */
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	/* (non-Javadoc)
-	 * @see java.lang.Object#hashCode()
-	 */
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + ((country == null) ? 0 : country.hashCode());
-		result = prime * result + ((formatted == null) ? 0 : formatted.hashCode());
-		result = prime * result + ((id == null) ? 0 : id.hashCode());
-		result = prime * result + ((locality == null) ? 0 : locality.hashCode());
-		result = prime * result + ((postalCode == null) ? 0 : postalCode.hashCode());
-		result = prime * result + ((region == null) ? 0 : region.hashCode());
-		result = prime * result + ((streetAddress == null) ? 0 : streetAddress.hashCode());
-		return result;
-	}
-
-	/* (non-Javadoc)
-	 * @see java.lang.Object#equals(java.lang.Object)
-	 */
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj) {
-			return true;
-		}
-		if (obj == null) {
-			return false;
-		}
-		if (!(obj instanceof Address)) {
-			return false;
-		}
-		Address other = (Address) obj;
-		if (country == null) {
-			if (other.country != null) {
-				return false;
-			}
-		} else if (!country.equals(other.country)) {
-			return false;
-		}
-		if (formatted == null) {
-			if (other.formatted != null) {
-				return false;
-			}
-		} else if (!formatted.equals(other.formatted)) {
-			return false;
-		}
-		if (id == null) {
-			if (other.id != null) {
-				return false;
-			}
-		} else if (!id.equals(other.id)) {
-			return false;
-		}
-		if (locality == null) {
-			if (other.locality != null) {
-				return false;
-			}
-		} else if (!locality.equals(other.locality)) {
-			return false;
-		}
-		if (postalCode == null) {
-			if (other.postalCode != null) {
-				return false;
-			}
-		} else if (!postalCode.equals(other.postalCode)) {
-			return false;
-		}
-		if (region == null) {
-			if (other.region != null) {
-				return false;
-			}
-		} else if (!region.equals(other.region)) {
-			return false;
-		}
-		if (streetAddress == null) {
-			if (other.streetAddress != null) {
-				return false;
-			}
-		} else if (!streetAddress.equals(other.streetAddress)) {
-			return false;
-		}
-		return true;
-	}
+	public void setCountry(String country);
 
 }
