+++ openid-connect-common/src/main/java/org/mitre/openid/connect/model/DefaultUserInfo.java
@@ -499,7 +499,7 @@ public class DefaultUserInfo implements UserInfo {
 
 		if (obj.has("address") && obj.get("address").isJsonObject()) {
 			JsonObject addr = obj.get("address").getAsJsonObject();
-			ui.setAddress(new Address());
+			ui.setAddress(new DefaultAddress());
 
 			ui.getAddress().setFormatted(nullSafeGetString(addr, "formatted"));
 			ui.getAddress().setStreetAddress(nullSafeGetString(addr, "street_address"));
